permuter: wave-2 sweep wins (5 fns, +2.74pp)
Structural-room permuter sweep across sub-95% in-scope units. All wins
A/B-verified via tools/ninja-locked rebuild + objdiff-cli, no neighbor
regressions in the affected TUs.

  ShiftedDotProduct    SndAnalysis        63.60 -> 64.97  (+1.37pp)
  IIR4PoleFilter ctor  IIRFilter          86.40 -> 87.17  (+0.77pp)
  ReadCounts           CharBonesSamples   61.50 -> 61.93  (+0.44pp)
  DisplayEvents        DisplayEvents      91.94 -> 92.04  (+0.10pp)
  MakeHSL              Color              93.79 -> 93.85  (+0.06pp)

BoxMap returned 7 candidates but all 0% (no decomp body yet — at-fixability 0.12).

+++ src/system/char/CharBonesSamples.cpp
@@ -463,7 +463,7 @@ void CharBonesSamples::ReadCounts(BinStream& bs, int i2){
     unsigned int i = 0;
     int tmp;
     int numTypesToRead = Min(7, i2);
-    for(; i < numTypesToRead; i++){
+    for(; (int)i < numTypesToRead; i++){
         bs >> mCounts[i];
     }
     for(int numTypesRead = i; numTypesRead < i2; numTypesRead++){
